Update saveAndExit button
