+++ src/widgets/Survey/ui/SurveyHeader.tsx
@@ -5,14 +5,9 @@ import { SurveyContext } from '../lib';
 
 import { SaveAndExitButton } from '~/features/SaveAssessmentAndExit';
 import { MultiInformantTooltip } from '~/features/TakeNow';
-import { ROUTES, Theme } from '~/shared/constants';
+import { Theme } from '~/shared/constants';
 import { AvatarBase, BaseProgressBar, Box, Text } from '~/shared/ui';
-import {
-  HourMinute,
-  isStringExist,
-  useCustomMediaQuery,
-  useCustomNavigation,
-} from '~/shared/utils';
+import { HourMinute, isStringExist, useCustomMediaQuery } from '~/shared/utils';
 
 type Props = {
   progress?: number;
@@ -25,18 +20,9 @@ const SurveyHeader = (props: Props) => {
   const context = useContext(SurveyContext);
 
   const { greaterThanSM } = useCustomMediaQuery();
-  const navigator = useCustomNavigation();
 
   const cutStringToLength = (str: string, length: number) => {
     return str.length > length ? `${str.substring(0, length)}...` : str;
-  };
-
-  const onSaveAndExitClick = () => {
-    return navigator.navigate(
-      context.publicAppletKey
-        ? ROUTES.publicJoin.navigateTo(context.publicAppletKey)
-        : ROUTES.appletDetails.navigateTo(context.appletId),
-    );
   };
 
   return (
@@ -78,7 +64,10 @@ const SurveyHeader = (props: Props) => {
             </Text>
           </Box>
           {!greaterThanSM && props.isSaveAndExitButtonShown && (
-            <SaveAndExitButton onClick={onSaveAndExitClick} />
+            <SaveAndExitButton
+              publicAppletKey={context.publicAppletKey}
+              appletId={context.appletId}
+            />
           )}
         </Box>
         {props.progress !== undefined && (
@@ -95,7 +84,10 @@ const SurveyHeader = (props: Props) => {
           justifyContent="center"
           justifySelf="flex-end"
         >
-          <SaveAndExitButton onClick={onSaveAndExitClick} />
+          <SaveAndExitButton
+            publicAppletKey={context.publicAppletKey}
+            appletId={context.appletId}
+          />
         </Box>
       )}
     </Box>
